pyproject.toml
[tool.poetry]
name = "lp-ft_ntk"
version = "0.1.0"
description = ""
authors = ["tom4649"]
readme = "README.md"
packages = [
    { include = "model" },
    { include = "tasks" },
    { include = "training" },
    { include = "utils" },
]

[tool.poetry.dependencies]
python = "^3.9"
adapter-transformers = "2.2.0"
torch = {path = "wheels/torch-1.12.1+cu113-cp39-cp39-linux_x86_64.whl"}
torchvision = {path = "wheels/torchvision-0.13.1+cu113-cp39-cp39-linux_x86_64.whl"}
aiohttp = "3.8.1"
aiosignal = "1.2.0"
albumentations = "0.4.3"
async-timeout = "4.0.1"
asynctest = "0.13.0"
attrs = "21.2.0"
audioread = "2.1.9"
backports-shutil-get-terminal-size = "1.0.0"
blessings = "1.7"
certifi = "2021.10.8"
cffi = "1.15.0"
charset-normalizer = "2.0.9"
click = "8.0.3"
configparser = "5.2.0"
configspace = "0.4.20"
cycler = "0.11.0"
cython = "0.29.26"
datasets = "1.15.1"
decorator = "5.1.0"
dill = "0.3.4"
docker-pycreds = "0.4.0"
docopt = "0.6.2"
filelock = "3.4.0"
fonttools = "4.28.5"
frozenlist = "1.2.0"
fsspec = "2021.11.1"
gitdb = "4.0.9"
gitpython = "3.1.24"
googledrivedownloader = "0.4"
gpustat = "0.6.0"
hpbandster = "0.7.4"
huggingface-hub = "0.2.1"
idna = "3.3"
imageio = "2.13.4"
imgaug = "0.2.6"
importlib-metadata = "4.8.2"
joblib = "1.1.0"
kiwisolver = "1.3.2"
librosa = "0.7.2"
matplotlib = "3.5.1"
multidict = "5.2.0"
multiprocess = "0.70.12.2"
netifaces = "0.11.0"
networkx = "2.6.3"
numba = "^0.58.1"
nvidia-ml-py3 = "7.352.0"
opencv-python = "4.5.4.60"
packaging = "21.3"
pandas = "1.3.4"
pathtools = "0.1.2"
patsy = "0.5.2"
pillow = "8.4.0"
pipreqs = "0.4.11"
promise = "2.3"
protobuf = "3.19.1"
psutil = "5.8.0"
pyarrow = "6.0.1"
pycparser = "2.21"
pyparsing = "3.0.6"
pyro4 = "4.81"
python-dateutil = "2.8.2"
pytz = "2021.3"
pywavelets = "1.2.0"
pyyaml = "6.0"
regex = "2021.11.10"
requests = "2.26.0"
resampy = "0.2.2"
sacremoses = "0.0.46"
scikit-image = "0.19.1"
scikit-learn = "1.0.1"
sentry-sdk = "1.5.0"
seqeval = "1.2.2"
serpent = "1.40"
shortuuid = "1.0.8"
six = "1.16.0"
smmap = "5.0.0"
soundfile = "0.10.3.post1"
statsmodels = "0.13.1"
subprocess32 = "3.5.4"
tabulate = "0.8.6"
tensorboardx = "2.0"
termcolor = "1.1.0"
threadpoolctl = "3.0.0"
tifffile = "2021.11.2"
tokenizers = "0.10.3"
tqdm = "4.62.3"
typing-extensions = "4.0.1"
urllib3 = "1.26.7"
wandb = "0.12.7"
xxhash = "2.0.2"
yarg = "0.1.9"
yarl = "1.7.2"
yaspin = "2.1.0"
zipp = "3.6.0"
hydra-core = "^1.3.2"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
